SLM-13 Move error codes into their own source file
